minor cleanup in response to Ben Jeurissen's recent changes to compile the Windows version

+++ cmd/mrflip.cpp
@@ -64,7 +64,7 @@ void run ()
 
   if (grad.is_set()) {
     Image::Transform transform (input_header);
-    for (uint dir = 0; dir < grad.rows(); dir++) {
+    for (size_t dir = 0; dir < grad.rows(); dir++) {
       Math::Vector<float> grad_flipped (3);
       transform.scanner2image_dir(grad.row(dir).sub(0,3), grad_flipped);
       grad_flipped[axis] = -grad_flipped[axis];
